requirements.txt
requests==2.32.3
praw==7.7.1
tweepy==4.14.0
Pillow==10.4.0
APScheduler==3.10.4
pytz==2024.1
python-dotenv==1.0.1